+++ src/main/java/com/wdelivery/member/controller/MemberController.java
@@ -63,41 +63,7 @@ public class MemberController {
 		return "main";
 	}
 
-	//회원정보 수정
-	@GetMapping("/mypageupdate.do")
-	public String mypageupdate(UserVO userVO, Model model, HttpSession session) {
-		
-		String user_email = (String) session.getAttribute("user_email"); //설정한 session 아이디
-		//System.out.println("mypage : " + user_email );
-		
-		userVO = memberService.userSelect(user_email); //세션 아이디 VO에 넣기
-		//System.out.println("mypage !!!!!!!!=>" + userVO.toString());
-		model.addAttribute("userVO", memberService.userSelect(userVO.getUser_email()));
-		
-		return "mypageupdate";
-	}
-	@PostMapping("/mypageUpdate.do")
-	public String mypageUpdate(UserVO userVO) {
-		//String user_email = (String) session.getAttribute("user_email");
-		//System.out.println("mypageupdate하는중 : " + user_email );
-		System.out.println("mypageupdateController" + userVO.getUser_seq());
-		//session.setAttribute("userVO", memberService.mypageUpdate(userVO));
-		memberService.mypageUpdate(userVO);
-		System.out.println(userVO.toString());
-		return "mypageupdate";
-	}
-	 
-	
 
-	@GetMapping("/addressBook.do")
-	public String addressBook() {
-		return "addressBook";
-	}
-
-	@GetMapping("/addressupdate.do")
-	public String addressUpdate() {
-		return "addressupdate";
-	}
 
 	@GetMapping("/order.do")
 	public String orderPage(Model model, @RequestParam(value="b_code", required=false) String b_code, 
@@ -213,11 +179,7 @@ public class MemberController {
 		return "join";
 	}
 
-	@GetMapping("/orderHistory.do")
-	public String orderHistory() {
-		return "orderHistory";
-	}
-
+	
 	//qnaselect
 	@PostMapping("/qnaSelect.do")
 	@ResponseBody
